daemon: Introduce the modular daemon application
This is the first step towards modularizing the initialization
of the cilium-agent. It converts runDaemon() into a uber/fx
application while retaining its structure and order of initialization.

For dumping the application dependency graph the option
"--dot-graph-output-file" was added. It will initialize the application
without starting and write out the graphviz dot file.

Signed-off-by: Jussi Maki <jussi@isovalent.com>