daemon: Only run syncEndpointsAndHostIPs via a controller
ReloadOnDeviceChange called into syncEndpointsAndHostIPs directly and
this might race with the call from controller (runs once a minute).
Refactor the code so that we only call this function via the controller.

Since UpdateController immediately runs the function there is no need to
call syncEndpointsAndHostIPs directly before it, though we do want to check
if the first sync fails.

Signed-off-by: Jussi Maki <jussi@isovalent.com>