status: Modularize status collector
This prepares the agent status collection to allow accepting probes
from outside daemon by allowing OnStatusUpdate to update the StatusResponse
directly via the 'StatusResponseUpdater' interface. The Daemon implements
this interface to keep the existing organization as-is.

Next steps with status would be to move the probes to their correct locations
and have the status package export the "GetStatus" API.

Signed-off-by: Jussi Maki <jussi@isovalent.com>