node: Move pkg/nodediscovery to pkg/node/nodediscovery
In preparation for refactoring of the node package to move ownership
of local node related information under the node discovery, move
the pkg/nodediscovery to pkg/node/nodediscovery.

Plan is to have pkg/node export an interface that provides access
to local node information and have that be implemented by pkg/node/nodediscovery
and thus makes more sense to locate the package under pkg/node.

Signed-off-by: Jussi Maki <jussi@isovalent.com>