promise: Add promise package
A simple implementation of promises. This can be used in situations where
a value is not available immediately and where there is multiple consumers
which rules out use of a simple channel.

The immediate use case will be to provide access to "*Daemon" and its fields
as we transition into a more modular construction.

Example usage:

type Foo struct { ... }

func getFoo() promise.Promise[*Foo] {
    resolve, p := promise.New[*Foo]()
    go func() {
	// long operation to construct *Foo
	resolve(&foo{...})
    }()
    return p
}

func useFoo(p promise.Promise[*Foo]) {
	foo := p.Await()
	...
}

Signed-off-by: Jussi Maki <jussi@isovalent.com>