hive: Add support for group of cells and decorators
This refactors the hive package in order to support "decorator" cells
that wrap a set of cells and provides them with augmented set of objects.

As it is looking more probable that it makes sense to implement directly on
top of dig, this commit drops use of fx.Option directly and instead implements
Provide and Invoke that take the functions directly. Additionally type aliases
are provided for fx.Hook and fx.Lifecycle. This way transitioning towards a lighter
implementation directly on top of dig is easier.

The cell related functions have been moved into "pkg/hive/cell" for cleaner
separation from hive functions. Changes to other packages other than pkg/hive
are to convert them to the new import path and shorter function names (NewCell->Provide).

Config is now a separate "cell type" and takes the default configuration as parameter.
This allows a cleaner way for providing defaults to a cell (see gops.Cell).

The configuration dumping in "cilium-agent objects" is for now removed as the
output was not very pretty and it is yet unclear how this should be done cleanly
with nested cells.

Signed-off-by: Jussi Maki <jussi@isovalent.com>