operator: Add leader lifecycle
Add a custom lifecycle to the operator hive that gets started when the
operator is elected leader. This allows implementing the operator functionality
as cells that gets started only after the operator becomes a leader and by
having the lifecycle conform to fx.Lifecycle we can have any cell use this lifecycle.

As first example the CEP watcher is now started with leader lifecycle.
To clean things up a bit the API server setup is separated into registerAPIServerHooks.

Signed-off-by: Jussi Maki <jussi@isovalent.com>