datapath: Implement DevicesController
DevicesController populates the devices and routes tables.

To avoid large refactoring of the existing code around option.Config.GetDevices()
this opts for an incremental approach: keep the DeviceManager API as-is and reuse
its test to validate that the existing functionality does not break.

The tests required some slight adaptation: the "DeviceManager" is re-created after
each test case to recreate the initial devices table.

Signed-off-by: Jussi Maki <jussi@isovalent.com>