datapath: Restart device controller on netlink failures
As the netlink protocol is unreliable (kernel can drop messages
on low-memory conditions), refactor the device controller to
flush state and restart the processing from scratch when any
of the netlink subscriptions fail or close unexpectedly.

The test suite reliability was improved by taking into account
the inherent eventually consistency of changes to the devices
table.

Signed-off-by: Jussi Maki <jussi@isovalent.com>