statedb2: Add experimental k8s reflector
Implement reflection of K8s objects onto a statedb table using the
client-go reflector.

With the Table[Obj].WorkQueue() we can implement the Resource[T] API
with statedb as the backing store.

Still to be done is providing the signal that the initial population
of the table is done (e.g. Resource[T] "Sync" event)

Signed-off-by: Jussi Maki <jussi@isovalent.com>