endpoint: Schedule sync-policy-map controller to retry ApplyPolicyMapChanges
The previous fix to avoid full dumping of policy map on every policy map change
(867e41a610) missed the likely accidental use of the sync-policy-map controller
for retrying failures in ApplyPolicyMapChanges.

To fix this, use the sync-policy-map controller both for retries of incremental
reconciliation and for the full periodic reconciliation.

Set a minimum retry wait time of 100 millis and a maximum of 1 minute. We want to
avoid aggressive retrying when the policy map is full.

Fixes: 867e41a610 ("endpoint: Only perform full synchronization periodically")
Signed-off-by: Jussi Maki <jussi@isovalent.com>