channels: Add channels package for common channel-related utilities
DoneChan: the "chan struct{}" pattern for signaling completion is used all over the place
and thus giving it a type alias can be beneficial in type signatures to underline
its usage.

ClosedDoneChan: when a "done channel" is used where nothing needs to be done, a new channel
needs to be created and immediately closed. This pattern seems to repeat a lot. We can instead
create a singleton DoneChan and close it at init time.

Signed-off-by: Jussi Maki <jussi@isovalent.com>