experimental: Improve the load-balancing models
Split off FrontendParams and BackendParams from Frontend and Backend.
This makes it explicit which fields are "user supplied" and which are not.

Introduce BackendInstance to carry per-service information about the backend
use. This allows a service to refer to a backend with a different port name and
assign a different weight to it.

For now remove the 'ID' fields as these can lead to confusion as they're filled
in during reconciliation. These were there only for help with debugging, e.g.
to match a Table[Frontend] entry with a BPF map key. At this point I think it is
cleaner to keep the internal state internal.

Signed-off-by: Jussi Maki <jussi@isovalent.com>