experimental: Wait for endpoints in K8s reflector before creating frontends
To avoid connection disruptions wait for endpoints before processing the service.
This way when we start with existing services and backends in the BPF maps we don't
temporarily overwrite the service entry with 0 backends.

There's still a slight chance that when there are multiple endpoint slices for a
service we might temporarily (e.g. for some seconds at most) reduce the number of
backends available for a service. TBD whether this is acceptable behavior.

Signed-off-by: Jussi Maki <jussi@isovalent.com>