watchers: Debounce service events
When a service has multiple endpoint slices associated with it
this will cause Cilium to process each slice separately, causing
repeated updates to the service BPF map and multiple hubble service
events.

To fix this, debounce the service events in the watcher before they're
processed.

The cilium/stream dependency was added to bring in the Buffer() operator.

Signed-off-by: Jussi Maki <jussi@isovalent.com>